First commit
Then God said, “Let the land produce vegetation: seed-bearing plants and
trees on the land that bear fruit with seed in it, according to their
various kinds.” And it was so. The land produced vegetation: plants
bearing seed according to their kinds and trees bearing fruit with seed
in it according to their kinds. And God saw that it was good.